Move report model to approprate location
